massive overhaul of everything for starting GPU support
this pushes the refactor of the matrix hierarchy through part of the
library. The part that compiles is enabled in the cmake; the part that
doesn't is disabled and can be added back incrementally. Sorry it's
such a big commit; sorry for the inconsistent style.

the enabled tests all build, run, and pass for me, except one. the
axpy test on multiple MPI ranks is broken.